requirements.txt
numpy>=1.21.0
torch>=2.0.0
umap-learn>=0.5.3
scikit-learn>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0